Now PathPayment classes have StrictReceive in their name to follow with the protocol 12 changes. Modified tests. Removed old files. Refactorized code.
This is a breaking changes commit.
