+++ stellar-dotnet-sdk/responses/results/OperationResult.cs
@@ -34,8 +34,8 @@ namespace stellar_dotnet_sdk.responses.results
                     return CreateAccountResult.FromXdr(result.CreateAccountResult);
                 case xdr.OperationType.OperationTypeEnum.PAYMENT:
                     return PaymentResult.FromXdr(result.PaymentResult);
-                case xdr.OperationType.OperationTypeEnum.PATH_PAYMENT:
-                    return PathPaymentResult.FromXdr(result.PathPaymentResult);
+                case xdr.OperationType.OperationTypeEnum.PATH_PAYMENT_STRICT_RECEIVE:
+                    return PathPaymentStrictReceiveResult.FromXdr(result.PathPaymentStrictReceiveResult);
                 case xdr.OperationType.OperationTypeEnum.MANAGE_BUY_OFFER:
                     return ManageBuyOfferResult.FromXdr(result.ManageBuyOfferResult);
                 case xdr.OperationType.OperationTypeEnum.MANAGE_SELL_OFFER:
