+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveMalformed.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Bad input.
     /// </summary>
-    public class PathPaymentMalformed : PathPaymentResult
+    public class PathPaymentStrictReceiveMalformed : PathPaymentStrictReceiveResult
     {
 
     }
