+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveNoDestination.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Destination account does not exist.
     /// </summary>
-    public class PathPaymentNoDestination : PathPaymentResult
+    public class PathPaymentStrictReceiveNoDestination : PathPaymentStrictReceiveResult
     {
 
     }
