+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveNoIssuer.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Missing issuer on one asset.
     /// </summary>
-    public class PathPaymentNoIssuer : PathPaymentResult
+    public class PathPaymentStrictReceiveNoIssuer : PathPaymentStrictReceiveResult
     {
         /// <summary>
         /// The asset that caused the error.
