+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveNoTrust.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Destination missing a trust line for asset.
     /// </summary>
-    public class PathPaymentNoTrust : PathPaymentResult
+    public class PathPaymentStrictReceiveNoTrust : PathPaymentStrictReceiveResult
     {
 
     }
