+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveNotAuthorized.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Destination not authorized to hold asset.
     /// </summary>
-    public class PathPaymentNotAuthorized : PathPaymentResult
+    public class PathPaymentStrictReceiveNotAuthorized : PathPaymentStrictReceiveResult
     {
 
     }
