+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveOfferCrossSelf.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Would cross one of its own offers.
     /// </summary>
-    public class PathPaymentOfferCrossSelf : PathPaymentResult
+    public class PathPaymentStrictReceiveOfferCrossSelf : PathPaymentStrictReceiveResult
     {
 
     }
