+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveOverSendmax.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Could not satisfy sendmax.
     /// </summary>
-    public class PathPaymentOverSendmax : PathPaymentResult
+    public class PathPaymentStrictReceiveOverSendmax : PathPaymentStrictReceiveResult
     {
 
     }
