+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveSrcNoTrust.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// No trust line on source account.
     /// </summary>
-    public class PathPaymentSrcNoTrust : PathPaymentResult
+    public class PathPaymentStrictReceiveSrcNoTrust : PathPaymentStrictReceiveResult
     {
 
     }
