+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveTooFewOffers.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Not enough offers to satisfy path.
     /// </summary>
-    public class PathPaymentTooFewOffers : PathPaymentResult
+    public class PathPaymentStrictReceiveTooFewOffers : PathPaymentStrictReceiveResult
     {
 
     }
