+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictReceiveUnderfunded.cs
@@ -3,7 +3,7 @@ namespace stellar_dotnet_sdk.responses.results
     /// <summary>
     /// Not enough funds in source account.
     /// </summary>
-    public class PathPaymentUnderfunded : PathPaymentResult
+    public class PathPaymentStrictReceiveUnderfunded : PathPaymentStrictReceiveResult
     {
 
     }
