+++ stellar-dotnet-sdk/responses/results/PathPaymentStrictSendResultOverSendmax.cs
@@ -1,0 +1,10 @@
+namespace stellar_dotnet_sdk.responses.results
+{
+    /// <summary>
+    /// Could not satisfy sendmax.
+    /// </summary>
+    public class PathPaymentStrictSendResultOverSendmax : PathPaymentStrictReceiveResult
+    {
+
+    }
+}
